safari-600.7-branch/LayoutTests/storage/indexeddb/noblobs-expected.txt

Confirm Blob/File/FileList limitations of WebKit's IndexedDB implementation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "noblobs.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
store = db.createObjectStore('storeName')
store.put('value', 'key')

testBlob():
PASS FileReader != null is true
test_content = 'This is a test. This is only a test.'
blob = new Blob([test_content])

validateExceptions(blob):
transaction = db.transaction('storeName', 'readwrite')
store = transaction.objectStore('storeName')
Expecting exception from store.put(blob, 'key')
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'put' on 'IDBObjectStore': The object store currently does not support blob values.
Expecting exception from store.add(blob, 'key')
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'add' on 'IDBObjectStore': The object store currently does not support blob values.
request = store.openCursor()
cursor = request.result
Expecting exception from cursor.update(blob)
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'update' on 'IDBCursor': The object store currently does not support blob values.

testFile():
file = fileInput.files[0]

validateExceptions(file):
transaction = db.transaction('storeName', 'readwrite')
store = transaction.objectStore('storeName')
Expecting exception from store.put(file, 'key')
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'put' on 'IDBObjectStore': The object store currently does not support blob values.
Expecting exception from store.add(file, 'key')
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'add' on 'IDBObjectStore': The object store currently does not support blob values.
request = store.openCursor()
cursor = request.result
Expecting exception from cursor.update(file)
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'update' on 'IDBCursor': The object store currently does not support blob values.

testFileList():
filelist = fileInput.files

validateExceptions(filelist):
transaction = db.transaction('storeName', 'readwrite')
store = transaction.objectStore('storeName')
Expecting exception from store.put(filelist, 'key')
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'put' on 'IDBObjectStore': The object store currently does not support blob values.
Expecting exception from store.add(filelist, 'key')
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'add' on 'IDBObjectStore': The object store currently does not support blob values.
request = store.openCursor()
cursor = request.result
Expecting exception from cursor.update(filelist)
PASS Exception was thrown.
PASS code is DOMException.DATA_CLONE_ERR
Exception message: Failed to execute 'update' on 'IDBCursor': The object store currently does not support blob values.
PASS successfullyParsed is true

TEST COMPLETE
